refactor: convert to module system
Code gets cleaner the more you split it up.
