+++ modules/index.js
@@ -1,3 +1,7 @@
+import 'https://unpkg.com/tesseract.js@2.0.0-beta.1/dist/tesseract.min.js';
+import colors from './variables.js';
+import convertResultsToTemplateCommands from './convertToWiki.js';
+
 const canvas = document.createElement('canvas');
 const output = document.getElementById('output');
 
@@ -117,7 +121,7 @@ const parseScriptImage = async (data) => {
 
       const details = collectTextDetails(ocrText);
 
-      if (entryType === 'loopStart') {
+      if (entryType === 'loopStart' && details[1]) {
         details[1].loopBreaks = doesLoopBreakOnError(entry);
         details[1].until = details[1].until || false;
       }
@@ -185,60 +189,6 @@ const parseScriptImage = async (data) => {
 
   // console.log('All images processed.');
 })();
-
-function convertResultsToTemplateCommands(results){
-  let template = '';
-
-  for (const entry of results) {
-    let value = false;
-    switch (entry.type) {
-      case 'loopStart':
-        template += `<br/>{{loop${getLoopType(entry)}`;
-        if (entry.details[1] && entry.details[1].loopBreaks) {
-          template += `|fail=`;
-        }
-        if (entry.details[1]) {
-          if (entry.details[1].until) {
-            if (entry.details[1].not) {
-              template += `|not=`
-            }
-            if (entry.details[1].empty) {
-              template += `|empty=`
-            }
-          }
-        }
-
-        template += `|commands=`
-        break;
-      case 'loopEnd':
-        template += `}}`
-        break;
-      case 'find':
-      case 'set':
-      case 'engage':
-      case 'move':
-      case 'add':
-      case 'take':
-      case 'pickup':
-        value = true;
-      case 'shout':
-      case 'use':
-      case 'disengage':
-      case 'recharge':
-      case 'cycleup':
-      case 'cycledown':
-      case 'retrieve':
-      case 'stow':
-      case 'swap':
-      case 'drop':
-      case 'wait':
-        template += `<br/>{{Command|type=${entry.type}${value?`|value=${entry.details[0]}`:''}}}`;
-        break;
-    }
-  }
-
-  return template;
-}
 
 function collectTextDetails(text) {
   if (text.startsWith('Move to')) return [text.replace('Move to ', ''), false];
@@ -528,18 +478,4 @@ function getPixelInEntry(x, y, entry) {
   const start = y * lineLength + (x * 4);
   const end = start + 4;
   return entry.slice(start, end);
-}
-
-function getLoopType(entry) {
-  const text = entry.details[0];
-  switch(text) {
-    default: return `|type=?|label=${text}`;
-    case 'forever!': return ''; // Loop template defaults to "forever"
-    case 'backpack':
-    case 'hands':
-    case 'held object': return `|type=${text}`;
-    case 'times':
-    case 'hear': return `|type=${text}|input=`; // TODO: input (eventually)
-    case 'until hear': return `|type=hear|input=`; // TODO: input (eventually)
-  }
 }
